+++ src/hfs/hfs_io.h
@@ -16,7 +16,7 @@ extern hfs_forktype_t HFSResourceForkType;
 
 #pragma mark HFS Volume
 
-ssize_t hfs_read            (void* buffer, const HFSPlus* hfs, size_t size, size_t offset) __attribute__((nonnull));
+ssize_t hfs_read            (void* buffer, const HFSPlus* hfs, size_t size, off_t offset) __attribute__((nonnull));
 ssize_t hfs_read_blocks     (void* buffer, const HFSPlus* hfs, size_t block_count, size_t start_block) __attribute__((nonnull));
 
 //FILE* fopen_hfs           (HFSPlus* hfs) __attribute__((nonnull));
@@ -29,7 +29,7 @@ int  hfsfork_make        (HFSPlusFork** fork, const HFSPlus* hfs, const HFSPlusF
 void hfsfork_free        (HFSPlusFork* fork) __attribute__((nonnull));
 
 ssize_t hfs_read_fork       (void* buffer, const HFSPlusFork* fork, size_t block_count, size_t start_block) __attribute__((nonnull));
-ssize_t hfs_read_fork_range (void* buffer, const HFSPlusFork* fork, size_t size, size_t offset) __attribute__((nonnull));
+ssize_t hfs_read_fork_range (void* buffer, const HFSPlusFork* fork, size_t size, off_t offset) __attribute__((nonnull));
 
 FILE* fopen_hfsfork       (HFSPlusFork* fork) __attribute__((nonnull));
